requirements.txt
# Web framework
fastapi>=0.100.0
uvicorn>=0.23.0
jinja2>=3.1.2
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0

# HTTP client
httpx>=0.24.0
requests>=2.31.0
firecrawl-py>=1.0.0

# PDF processing
PyMuPDF>=1.23.0
pikepdf>=8.0.0
pdfplumber>=0.10.0
pdfminer.six>=20221105
pdf2image>=1.16.0
Pillow>=10.0.0
numpy>=1.24.0

# Image processing for visual diffs
opencv-python-headless>=4.8.0

# AWS (for OCR)
boto3>=1.28.0

# Configuration
python-dotenv>=1.0.0

# Logging
structlog>=23.1.0

# Flask for test server
flask>=3.0.0
